fix(ui): remove title id and handle error catch

+++ src/ui/components/PasswordModule/PasswordModule.tsx
@@ -16,8 +16,10 @@ import { PasswordModuleProps, PasswordModuleRef } from "./PasswordModule.types";
 import {
   getStateCache,
   setAuthentication,
+  setToastMsg,
 } from "../../../store/reducers/stateCache";
 import { useAppDispatch } from "../../../store/hooks";
+import { ToastMsgType } from "../../globals/types";
 
 const PasswordModule = forwardRef<PasswordModuleRef, PasswordModuleProps>(
   ({ title, isOnboarding, description, testId, onCreateSuccess }, ref) => {
@@ -120,11 +122,19 @@ const PasswordModule = forwardRef<PasswordModuleRef, PasswordModuleProps>(
           );
         } else {
           try {
-            await Agent.agent.basicStorage.deleteById(
-              MiscRecordId.OP_PASS_HINT
-            );
+            const previousHint = (
+              await Agent.agent.basicStorage.findById(MiscRecordId.OP_PASS_HINT)
+            )?.content?.value;
+
+            if (previousHint) {
+              await Agent.agent.basicStorage.deleteById(
+                MiscRecordId.OP_PASS_HINT
+              );
+            }
           } catch (e) {
-            // TODO: handle error
+            // eslint-disable-next-line no-console
+            console.error("Unable to delete password hint", e);
+            dispatch(setToastMsg(ToastMsgType.UNABLE_DELETE_PASSWORD_HINT));
           }
         }
       }
